* extract p2p_pipeline definition from the OP definition
